Fixes for reload of profilepage and show of worldranking. Moved some shared functions to new class Fix for not showing toast multiple times when not needed
